[Fix] RecipeTags table => 外部キーをbigint型へ修正

+++ db/migrate/20210510165017_change_column_to_recipe_tags.rb
@@ -1,12 +1,12 @@
 class ChangeColumnToRecipeTags < ActiveRecord::Migration[5.2]
 
   def up
-    change_column :recipe_tags, :recipe_id, :integer
-    change_column :recipe_tags, :tag_id, :integer
+    change_column :recipe_tags, :recipe_id, :bigint
+    change_column :recipe_tags, :tag_id, :bigint
   end
 
   def down
-    change_column :recipe_tags, :recipe_id, :bigint
-    change_column :recipe_tags, :tag_id, :bigint
+    change_column :recipe_tags, :recipe_id, :integer
+    change_column :recipe_tags, :tag_id, :integer
   end
 end
